# spec/api/openapi.yaml
openapi: 3.2.1
...
info:
  title: Knowledge Platform Document API
  version: 0.1.0
  description: One authorized Document API for human, agent and CLI clients. Identity is established only by a trusted
    adapter; request data never supplies the actor.
servers:
- url: /
  description: Same-origin deployment; no production endpoint is fixed.
...
paths:
  /v1/documents:
    get:
      tags:
      - Documents
      summary: List documents by authorized view
      operationId: listDocuments
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/DocumentList'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - $ref: '#/components/parameters/view'
      - $ref: '#/components/parameters/titleContains'
      - $ref: '#/components/parameters/queryFolderId'
      - $ref: '#/components/parameters/includeDescendants'
      - $ref: '#/components/parameters/documentType'
      - $ref: '#/components/parameters/owningDepartment'
      - $ref: '#/components/parameters/category'
      - $ref: '#/components/parameters/createdFrom'
      - $ref: '#/components/parameters/createdBefore'
      - $ref: '#/components/parameters/sort'
      - $ref: '#/components/parameters/pageSize'
      - $ref: '#/components/parameters/cursor'
      - $ref: '#/components/parameters/unreadOnly'
    post:
      tags:
      - Documents
      summary: Create initial document and primary file
      operationId: createDocument
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '201':
          $ref: '#/components/responses/CreateDocumentResult'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        $ref: '#/components/requestBodies/CreateDocumentMultipart'
  /v1/document-creation-outcomes/{documentId}:
    parameters:
    - $ref: '#/components/parameters/documentId'
    get:
      tags:
      - Documents
      summary: Recover an initial document creation outcome
      operationId: recoverDocumentCreation
      description: All generated IDs must match and current authoring authorization is re-evaluated before disclosure.
      parameters:
      - name: documentVersionId
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - name: fileId
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          $ref: '#/components/responses/CreateDocumentResult'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/documents/{documentId}/versions/{versionId}/files:
    parameters:
    - $ref: '#/components/parameters/documentId'
    - $ref: '#/components/parameters/versionId'
    get:
      tags:
      - Files
      summary: List authorized version files
      operationId: listVersionFiles
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/FileList'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - $ref: '#/components/parameters/purpose'
  /v1/documents/{documentId}/versions/{versionId}/files/{contentItemId}/{representationId}:
    parameters:
    - $ref: '#/components/parameters/documentId'
    - $ref: '#/components/parameters/versionId'
    - $ref: '#/components/parameters/contentItemId'
    - $ref: '#/components/parameters/representationId'
    get:
      tags:
      - Files
      summary: Download only after authorization and Audit commit
      operationId: downloadVersionFile
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/BinaryDownload'
        '401':
          $ref: '#/components/responses/Problem'
        '416':
          description: Range requests are not supported in v0; no file bytes are returned.
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - $ref: '#/components/parameters/purpose'
...
components:
  schemas:
    Uuid:
      $ref: ./schemas/common/identifiers.yaml#/$defs/Uuid
    Cursor:
      $ref: ./schemas/common/identifiers.yaml#/$defs/Cursor
    FieldError:
      $ref: ./schemas/errors/field-error.yaml
    Problem:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      - traceId
      - retryable
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Stable machine code from spec/errors/error-registry.yaml.
        traceId:
          type: string
        retryable:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        recovery:
          $ref: '#/components/schemas/CreateRecovery'
        exactRetry:
          type: boolean
          description: Only the same operation ID and payload may be retried.
    CreateRecovery:
      type: object
      additionalProperties: false
      required:
      - documentId
      - documentVersionId
      - fileId
      - recoveryEndpoint
      properties:
        documentId:
          $ref: '#/components/schemas/Uuid'
        documentVersionId:
          $ref: '#/components/schemas/Uuid'
        fileId:
          $ref: '#/components/schemas/Uuid'
        recoveryEndpoint:
          type: string
          format: uri-reference
...
    DocumentList:
      $ref: ./schemas/document/models.yaml#/$defs/DocumentList
...
    FileList:
      $ref: ./schemas/document/models.yaml#/$defs/FileList
...
    CreateDocumentResult:
      $ref: ./schemas/document/models.yaml#/$defs/CreateDocumentResult
...
    CreateDocument:
      $ref: ./schemas/document/commands.yaml#/$defs/CreateDocument
...
  parameters:
    documentId:
      name: documentId
      in: path
      required: true
      description: Document identifier.
      schema:
        $ref: '#/components/schemas/Uuid'
    versionId:
      name: versionId
      in: path
      required: true
      description: Version identifier.
      schema:
        $ref: '#/components/schemas/Uuid'
...
    contentItemId:
      name: contentItemId
      in: path
      required: true
      description: Content item identifier.
      schema:
        $ref: '#/components/schemas/Uuid'
    representationId:
      name: representationId
      in: path
      required: true
      description: Representation identifier.
      schema:
        $ref: '#/components/schemas/Uuid'
    view:
      name: view
      in: query
      required: true
      description: A distinct authorization and projection view.
      schema:
        type: string
        enum:
        - published
        - authoring
        - history
...
    purpose:
      name: purpose
      in: query
      required: true
      description: Authorization purpose; it cannot be omitted to broaden access.
      schema:
        type: string
        enum:
        - published
        - authoring
        - history
    titleContains:
      name: titleContains
      in: query
      required: false
      description: Title filter.
      schema:
        type: string
    queryFolderId:
      name: folderId
      in: query
      required: false
      description: Folder filter.
      schema:
        $ref: '#/components/schemas/Uuid'
    includeDescendants:
      name: includeDescendants
      in: query
      required: false
      description: Include visible descendants.
      schema:
        type: boolean
    documentType:
      name: documentType
      in: query
      required: false
      description: Document type filter.
      schema:
        type: string
    owningDepartment:
      name: owningDepartment
      in: query
      required: false
      description: Owning department filter.
      schema:
        type: string
    category:
      name: category
      in: query
      required: false
      description: Category filter.
      schema:
        type: string
    createdFrom:
      name: createdFrom
      in: query
      required: false
      description: Creation lower bound.
      schema:
        type: string
        format: date-time
    createdBefore:
      name: createdBefore
      in: query
      required: false
      description: Creation upper bound.
      schema:
        type: string
        format: date-time
    sort:
      name: sort
      in: query
      required: false
      description: Stable sort key.
      schema:
        type: string
    pageSize:
      name: pageSize
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    cursor:
      name: cursor
      in: query
      required: false
      description: Opaque, principal and query bound cursor.
      schema:
        $ref: '#/components/schemas/Cursor'
    unreadOnly:
      name: unreadOnly
      in: query
      required: false
      description: Valid only when view=published.
      schema:
        type: boolean
  requestBodies:
...
    CreateDocumentMultipart:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            additionalProperties: false
            required:
            - request
            - file
            properties:
              request:
                $ref: '#/components/schemas/CreateDocument'
              file:
                type: string
                format: binary
          encoding:
            request:
              contentType: application/json
            file:
              contentType: application/octet-stream
...
  responses:
    DocumentList:
      description: DocumentList
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DocumentList'
          example:
            view: published
            items: []
            nextCursor: null
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema:
            type: string
            const: private, no-store
...
    FileList:
      description: FileList
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FileList'
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema:
            type: string
            const: private, no-store
...
    CreateDocumentResult:
      description: CreateDocumentResult
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateDocumentResult'
          example:
            documentId: 00000000-0000-4000-8000-000000000001
            documentVersionId: 00000000-0000-4000-8000-000000000002
            fileId: 00000000-0000-4000-8000-000000000003
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema:
            type: string
            const: private, no-store
...
    BinaryDownload:
      description: Authorized and audited full file body; Range is rejected in v0.
      headers:
        Content-Disposition:
          schema:
            type: string
          description: attachment with sanitized safe_display_name
        Content-Length:
          schema:
            type: integer
            minimum: 0
        X-Content-Type-Options:
          schema:
            type: string
            const: nosniff
        Cache-Control:
          schema:
            type: string
            const: private, no-store
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
    Problem:
      description: RFC 9457 problem with stable machine code.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Validation failed
            status: 422
            code: VALIDATION_FAILED
            traceId: example-trace
            retryable: false
            errors: []
      headers:
        Cache-Control:
          schema:
            type: string
            const: private, no-store